Move `httpcache` SQL code to sql.go
